+++ src/applications/personalization/profile360/vet360/constants/index.js
@@ -81,3 +81,10 @@ export const API_ROUTES = {
   EMAILS: '/profile/email_addresses',
   ADDRESSES: '/profile/addresses'
 };
+
+export const VET360_INITIALIZATION_STATUS = {
+  INITIALIZED: 'INITIALIZED',
+  INITIALIZING: 'INITIALIZING',
+  INITIALIZATION_FAILURE: 'INITIALIZATION_FAILURE',
+  UNINITALIZED: 'UNINITALIZED'
+};
